Add a reader/walker for RISOT corpus.
Add a walker which will walk the RISOT corpus to return all the files present
in the corpus. It can return origianl and ocred versions of files or both at
the same time.

Add doccumentation for it.

Initialis the strucutre for documentation of api and include it in the table of
contents.